finished refactoring static component
