+++ src/components/resume/Resume.js
@@ -2,6 +2,7 @@ import React from "react";
 import { useParams } from "react-router-dom";
 import { theme } from "../../theme/trbTheme";
 import { Box, Flex } from "rebass";
+import { ProjectsList } from "../posts/projects/ProjectsList";
 import { SkillsList } from "../posts/skills/SkillsList";
 import { ExpsList } from "../posts/exps/ExpsList";
 
@@ -9,11 +10,7 @@ export const Resume = () => {
   const { userId } = useParams();
   return (
     <Container>
-      <ProjectsContainer>
-        <ProjectBox>project 1</ProjectBox>
-        <ProjectBox>project 2</ProjectBox>
-        <ProjectBox>project 3</ProjectBox>
-      </ProjectsContainer>
+      <ProjectsList />
       <VerticalSplit>
         <SkillsList userId={userId} />
         <ExpsList userId={userId} />
@@ -45,38 +42,6 @@ const VerticalSplit = (props) => (
       flexBasis: "60%",
       flexDirection: "row",
       justifyContent: "space-between",
-    }}
-  />
-);
-
-const ProjectsContainer = (props) => (
-  <Flex
-    {...props}
-    theme={theme}
-    width="90%"
-    sx={{
-      flexDirection: "row",
-      flexWrap: "wrap",
-      flexGrow: "1",
-      flexShrink: "1",
-      flexBasis: "30%",
-      justifyContent: "space-between",
-    }}
-  />
-);
-
-const ProjectBox = (props) => (
-  <Box
-    {...props}
-    theme={theme}
-    minHeight="10rem"
-    m="0.5rem"
-    textAlign="center"
-    bg="azure"
-    sx={{
-      flexBasis: "30%",
-      flexGrow: "1",
-      flexShrink: "1",
     }}
   />
 );
